refactor onValidateSetup
