+++ src/actions/index.ts
@@ -1,5 +1,6 @@
 export { default as onStartup } from './onStartup'
 export { default as onTutorialConfig } from './onTutorialConfig'
 export { default as onTutorialContinueConfig } from './onTutorialContinueConfig'
+export { default as onValidateSetup } from './onValidateSetup'
 export { default as onErrorPage } from './onErrorPage'
 export { default as onTestPass } from './onTestPass'
